Experimenting with a finale idea
During panic stages, supply balloons will appear and drop cash packages. This is to provide a source of currency during a finale and to encourage players to take risks. This should also create interesting scenarios where players must provide cover fire to whoever is unpacking the currency.

Ultimately I couldn't figure out a way to detect when panic stages have started via a map script.
So this idea is now on hold.