Persistence between maps. But it's scuffed...
The default saving and restoring systems are incredibly scuffed. There are so many weird bugs that alter the data you pass in.
This system is so broken and creates so many problems and doesn't even prevent data persistence between different campagins (ironically the documentation says that this won't happen).
Thus, time to redo everything and create my own system.